pnsv_workspace/requirements.txt
pydantic>=2.0
dspy-ai
openai
pytest
